requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.4 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
anyio==3.6.2 ; python_full_version >= '3.6.2'
appnope==0.1.3 ; platform_system == 'Darwin'
asttokens==2.2.1
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==22.2.0 ; python_version >= '3.6'
backcall==0.2.0
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
click==8.1.3 ; python_version >= '3.7'
comm==0.1.3 ; python_version >= '3.6'
contourpy==1.0.7 ; python_version >= '3.8'
cycler==0.11.0 ; python_version >= '3.6'
debugpy==1.6.6 ; python_version >= '3.7'
decorator==5.1.1 ; python_version >= '3.5'
dnspython==2.3.0 ; python_version >= '3.7' and python_version < '4.0'
email-validator==1.3.1
executing==1.2.0
fastapi[all]==0.95.0
flask==2.2.3
fonttools==4.39.3 ; python_version >= '3.8'
frozenlist==1.3.3 ; python_version >= '3.7'
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.16.3 ; python_version >= '3.7'
httptools==0.5.0
httpx==0.23.3 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
ipykernel==6.22.0
ipython==8.12.0 ; python_version >= '3.8'
ipywidgets==8.0.6
itsdangerous==2.1.2 ; python_version >= '3.7'
jedi==0.18.2 ; python_version >= '3.6'
jinja2==3.1.2 ; python_version >= '3.7'
joblib==1.2.0 ; python_version >= '3.7'
jupyter-client==8.1.0 ; python_version >= '3.8'
jupyter-core==5.3.0 ; python_version >= '3.8'
jupyterlab-widgets==3.0.7 ; python_version >= '3.7'
kiwisolver==1.4.4 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
matplotlib==3.7.1
matplotlib-inline==0.1.6 ; python_version >= '3.5'
multidict==6.0.4 ; python_version >= '3.7'
nest-asyncio==1.5.6 ; python_version >= '3.5'
numpy==1.24.2
openai==0.27.2
orjson==3.8.9
packaging==23.0 ; python_version >= '3.7'
pandas==1.5.3
parso==0.8.3 ; python_version >= '3.6'
pexpect==4.8.0 ; sys_platform != 'win32'
pickleshare==0.7.5
pillow==9.4.0 ; python_version >= '3.7'
platformdirs==3.2.0 ; python_version >= '3.7'
prompt-toolkit==3.0.38 ; python_full_version >= '3.7.0'
psutil==5.9.4 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.7 ; python_version >= '3.7'
pyee==9.0.4
pygments==2.14.0 ; python_version >= '3.6'
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pyyaml==6.0
pyzmq==25.0.2 ; python_version >= '3.6'
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
rfc3986[idna2008]==1.5.0
scikit-learn==1.2.2
scipy==1.10.1 ; python_version < '3.12' and python_version >= '3.8'
seaborn==0.12.2
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
slack-bolt==1.17.0
slack-sdk==3.20.2
slackeventsapi==3.0.1
sniffio==1.3.0 ; python_version >= '3.7'
stack-data==0.6.2
starlette==0.26.1 ; python_version >= '3.7'
tabulate==0.9.0
threadpoolctl==3.1.0 ; python_version >= '3.6'
tornado==6.2 ; python_version >= '3.7'
tqdm==4.65.0
traitlets==5.9.0 ; python_version >= '3.7'
typing-extensions==4.5.0 ; python_version >= '3.7'
ujson==5.7.0
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn[standard]==0.21.1 ; python_version >= '3.7'
uvloop==0.17.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==10.4 ; python_version >= '3.7'
werkzeug==2.2.3 ; python_version >= '3.7'
widgetsnbextension==4.0.7 ; python_version >= '3.7'
yarl==1.8.2 ; python_version >= '3.7'
